+++ src/test/java/net/eightlives/friendlyssl/integration/tests/AuthChallengeTimeoutTest.java
@@ -1,5 +1,7 @@
-package net.eightlives.friendlyssl.integration;
+package net.eightlives.friendlyssl.integration.tests;
 
+import net.eightlives.friendlyssl.integration.IntegrationTest;
+import net.eightlives.friendlyssl.integration.TestApp;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
@@ -13,18 +15,20 @@ import org.springframework.test.context.ContextConfiguration;
 import org.testcontainers.Testcontainers;
 import org.testcontainers.containers.BindMode;
 import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.junit.jupiter.Container;
 
 import java.util.List;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@ContextConfiguration(initializers = OrderTimeoutTest.class, classes = TestApp.class)
-@ActiveProfiles({"test-base", "test-order-timeout"})
-class OrderTimeoutTest implements IntegrationTest {
+@ContextConfiguration(initializers = AuthChallengeTimeoutTest.class, classes = TestApp.class)
+@ActiveProfiles({"test-base", "test-auth-challenge-timeout"})
+class AuthChallengeTimeoutTest implements IntegrationTest {
 
     static {
         Testcontainers.exposeHostPorts(5002);
     }
 
+    @Container
     static GenericContainer pebbleContainer = new GenericContainer("letsencrypt/pebble")
             .withCommand("pebble -config /test/my-pebble-config.json")
             .withExposedPorts(14000, 15000)
@@ -41,19 +45,19 @@ class OrderTimeoutTest implements IntegrationTest {
         return pebbleContainer;
     }
 
-    @DisplayName("Start server and certificate order does not respond as valid in time")
+    @DisplayName("Start server and and auth challenges do not respond as valid in time")
     @Timeout(20)
     @ExtendWith(OutputCaptureExtension.class)
     @DirtiesContext
     @Test
-    void orderTimeout(CapturedOutput output) {
+    void authChallengeTimeout(CapturedOutput output) {
         testLogOutput(
                 List.of(
                         "n.e.f.s.SSLCertificateCreateRenewService : Starting certificate create/renew",
                         "n.e.f.service.AcmeAccountService         : Account does not exist. Creating account.",
                         "n.e.f.s.SSLCertificateCreateRenewService : Certificate account login accessed",
                         "n.e.f.s.SSLCertificateCreateRenewService : Beginning certificate order. Renewal: false",
-                        "n.e.f.service.UpdateCheckerService       : Resource is valid",
+                        "n.e.f.l.ChallengeTokenRequestedListener  : Timeout while checking for challenge status",
                         "n.e.f.s.SSLCertificateCreateRenewService : Exception while ordering certificate, retry in 1 hours",
                         "net.eightlives.friendlyssl.exception.SSLCertificateException: Exception while handling SSL certificate management",
                         "Caused by: java.util.concurrent.TimeoutException: null"
@@ -62,3 +66,4 @@ class OrderTimeoutTest implements IntegrationTest {
         );
     }
 }
+
