+++ src/test/java/net/eightlives/friendlyssl/integration/tests/CertificateAutoRenewalTest.java
@@ -1,12 +1,22 @@
-package net.eightlives.friendlyssl.integration;
+package net.eightlives.friendlyssl.integration.tests;
 
+import net.eightlives.friendlyssl.annotation.FriendlySSL;
 import net.eightlives.friendlyssl.config.FriendlySSLConfig;
-import org.junit.jupiter.api.*;
+import net.eightlives.friendlyssl.integration.IntegrationTest;
+import net.eightlives.friendlyssl.util.TestConstants;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.system.CapturedOutput;
 import org.springframework.boot.test.system.OutputCaptureExtension;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Primary;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
@@ -14,18 +24,26 @@ import org.testcontainers.Testcontainers;
 import org.testcontainers.containers.BindMode;
 import org.testcontainers.containers.GenericContainer;
 
-import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.time.Clock;
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 
+import static net.eightlives.friendlyssl.integration.tests.CertificateAutoRenewalTest.CertificateRenewalApp;
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.when;
 
+/**
+ * Pebble cannot currently return a user-defined expiration time (https://github.com/letsencrypt/pebble/issues/251),
+ * but when it does another test similar to this one should be created that doesn't use a mock Clock.
+ */
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@ContextConfiguration(initializers = AccountReuseTest.class, classes = TestApp.class)
-@ActiveProfiles({"test-base", "test-tos-accepted"})
-@TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-class AccountReuseTest implements IntegrationTest {
+@ContextConfiguration(initializers = CertificateAutoRenewalTest.class, classes = CertificateRenewalApp.class)
+@ActiveProfiles({"test-base", "test-tos-accepted", "test-existing-keystore", "test-override-beans"})
+@ExtendWith(MockitoExtension.class)
+class CertificateAutoRenewalTest implements IntegrationTest {
 
     static {
         Testcontainers.exposeHostPorts(5002);
@@ -44,50 +62,51 @@ class AccountReuseTest implements IntegrationTest {
 
     @Autowired
     FriendlySSLConfig config;
+    @Mock
+    static Clock clock;
 
     @Override
     public GenericContainer getPebbleContainer() {
         return pebbleContainer;
     }
 
-    @Order(1)
-    @DisplayName("Start server and account does not exist")
+    @FriendlySSL
+    @SpringBootApplication
+    static class CertificateRenewalApp {
+
+        @Bean
+        @Primary
+        public Clock clock(FriendlySSLConfig config) {
+            Instant renewThreshold = TestConstants.EXISTING_KEYSTORE_CERT_EXPIRATION
+                    .minus(config.getAutoRenewalHoursBefore(), ChronoUnit.HOURS);
+            Instant oneSecondBeforeRenewThreshold = renewThreshold
+                    .minus(1, ChronoUnit.SECONDS);
+
+            when(clock.instant())
+                    .thenReturn(oneSecondBeforeRenewThreshold)
+                    .thenReturn(oneSecondBeforeRenewThreshold)
+                    .thenReturn(renewThreshold);
+
+            return clock;
+        }
+    }
+
+    @DisplayName("Server should auto-renew when auto-renew hours threshold time is crossed")
     @Timeout(20)
     @ExtendWith(OutputCaptureExtension.class)
     @DirtiesContext
     @Test
-    void accountDoesNotExist(CapturedOutput output) throws IOException {
-        testLogOutput(
+    void autoRenew(CapturedOutput output) {
+        testLogOutputExact(
                 List.of(
                         "n.e.f.s.SSLCertificateCreateRenewService : Starting certificate create/renew",
                         "n.e.f.service.AcmeAccountService         : Account does not exist. Creating account.",
                         "n.e.f.s.SSLCertificateCreateRenewService : Certificate account login accessed",
-                        "n.e.f.s.SSLCertificateCreateRenewService : Beginning certificate order. Renewal: false",
-                        "n.e.f.service.UpdateCheckerService       : Resource is valid",
-                        "n.e.f.s.SSLCertificateCreateRenewService : Certificate renewal successful. New certificate expiration time is"
-                ),
-                output
-        );
-
-        assertTrue(Files.exists(Path.of(config.getKeystoreFile())));
-
-        Files.delete(Path.of(config.getKeystoreFile()));
-    }
-
-    @Order(2)
-    @DisplayName("Then start server and account exists")
-    @Timeout(20)
-    @ExtendWith(OutputCaptureExtension.class)
-    @DirtiesContext
-    @Test
-    void accountExists(CapturedOutput output) {
-        testLogOutput(
-                List.of(
+                        "n.e.f.s.SSLCertificateCreateRenewService : Existing certificate expiration time is 2012-12-22T07:41:51Z",
                         "n.e.f.s.SSLCertificateCreateRenewService : Starting certificate create/renew",
                         "n.e.f.service.AcmeAccountService         : Using existing account login",
                         "n.e.f.s.SSLCertificateCreateRenewService : Certificate account login accessed",
-                        "n.e.f.s.SSLCertificateCreateRenewService : Beginning certificate order. Renewal: false",
-                        "n.e.f.service.UpdateCheckerService       : Resource is valid",
+                        "n.e.f.s.SSLCertificateCreateRenewService : Beginning certificate order. Renewal: true",
                         "n.e.f.s.SSLCertificateCreateRenewService : Certificate renewal successful. New certificate expiration time is"
                 ),
                 output
