+++ src/test/java/net/eightlives/friendlyssl/integration/tests/OrderTimeoutTest.java
@@ -1,5 +1,7 @@
-package net.eightlives.friendlyssl.integration;
+package net.eightlives.friendlyssl.integration.tests;
 
+import net.eightlives.friendlyssl.integration.IntegrationTest;
+import net.eightlives.friendlyssl.integration.TestApp;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
@@ -17,9 +19,9 @@ import org.testcontainers.containers.GenericContainer;
 import java.util.List;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@ContextConfiguration(initializers = SlowServerTest.class, classes = TestApp.class)
-@ActiveProfiles({"test-base", "test-tos-accepted"})
-class SlowServerTest implements IntegrationTest {
+@ContextConfiguration(initializers = OrderTimeoutTest.class, classes = TestApp.class)
+@ActiveProfiles({"test-base", "test-order-timeout"})
+class OrderTimeoutTest implements IntegrationTest {
 
     static {
         Testcontainers.exposeHostPorts(5002);
@@ -28,7 +30,7 @@ class SlowServerTest implements IntegrationTest {
     static GenericContainer pebbleContainer = new GenericContainer("letsencrypt/pebble")
             .withCommand("pebble -config /test/my-pebble-config.json")
             .withExposedPorts(14000, 15000)
-            .withEnv("PEBBLE_VA_NOSLEEP", "0")
+            .withEnv("PEBBLE_VA_NOSLEEP", "1")
             .withEnv("PEBBLE_WFE_NONCEREJECT", "0")
             .withClasspathResourceMapping(
                     "pebble-config.json",
@@ -41,12 +43,12 @@ class SlowServerTest implements IntegrationTest {
         return pebbleContainer;
     }
 
-    @DisplayName("Start server and ACME server responses have variable latency")
+    @DisplayName("Start server and certificate order does not respond as valid in time")
     @Timeout(20)
     @ExtendWith(OutputCaptureExtension.class)
     @DirtiesContext
     @Test
-    void slowServer(CapturedOutput output) {
+    void orderTimeout(CapturedOutput output) {
         testLogOutput(
                 List.of(
                         "n.e.f.s.SSLCertificateCreateRenewService : Starting certificate create/renew",
@@ -54,7 +56,9 @@ class SlowServerTest implements IntegrationTest {
                         "n.e.f.s.SSLCertificateCreateRenewService : Certificate account login accessed",
                         "n.e.f.s.SSLCertificateCreateRenewService : Beginning certificate order. Renewal: false",
                         "n.e.f.service.UpdateCheckerService       : Resource is valid",
-                        "n.e.f.s.SSLCertificateCreateRenewService : Certificate renewal successful. New certificate expiration time is"
+                        "n.e.f.s.SSLCertificateCreateRenewService : Exception while ordering certificate, retry in 1 hours",
+                        "net.eightlives.friendlyssl.exception.SSLCertificateException: Exception while handling SSL certificate management",
+                        "Caused by: java.util.concurrent.TimeoutException: null"
                 ),
                 output
         );
