+++ src/test/java/net/eightlives/friendlyssl/integration/tests/PartialNonceRejectTest.java
@@ -1,9 +1,13 @@
-package net.eightlives.friendlyssl.integration;
+package net.eightlives.friendlyssl.integration.tests;
 
+import net.eightlives.friendlyssl.config.FriendlySSLConfig;
+import net.eightlives.friendlyssl.integration.IntegrationTest;
+import net.eightlives.friendlyssl.integration.TestApp;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.system.CapturedOutput;
 import org.springframework.boot.test.system.OutputCaptureExtension;
@@ -13,55 +17,59 @@ import org.springframework.test.context.ContextConfiguration;
 import org.testcontainers.Testcontainers;
 import org.testcontainers.containers.BindMode;
 import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.junit.jupiter.Container;
 
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.List;
 
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
-@ContextConfiguration(initializers = AuthChallengeTimeoutTest.class, classes = TestApp.class)
-@ActiveProfiles({"test-base", "test-auth-challenge-timeout"})
-class AuthChallengeTimeoutTest implements IntegrationTest {
+@ContextConfiguration(initializers = PartialNonceRejectTest.class, classes = TestApp.class)
+@ActiveProfiles({"test-base", "test-tos-accepted"})
+class PartialNonceRejectTest implements IntegrationTest {
 
     static {
         Testcontainers.exposeHostPorts(5002);
     }
 
-    @Container
     static GenericContainer pebbleContainer = new GenericContainer("letsencrypt/pebble")
             .withCommand("pebble -config /test/my-pebble-config.json")
             .withExposedPorts(14000, 15000)
             .withEnv("PEBBLE_VA_NOSLEEP", "1")
-            .withEnv("PEBBLE_WFE_NONCEREJECT", "0")
+            .withEnv("PEBBLE_WFE_NONCEREJECT", "50")
             .withClasspathResourceMapping(
                     "pebble-config.json",
                     "/test/my-pebble-config.json",
                     BindMode.READ_ONLY
             );
 
+    @Autowired
+    FriendlySSLConfig config;
+
     @Override
     public GenericContainer getPebbleContainer() {
         return pebbleContainer;
     }
 
-    @DisplayName("Start server and and auth challenges do not respond as valid in time")
+    @DisplayName("Start server and certificate 50% of nonces are rejected")
     @Timeout(20)
     @ExtendWith(OutputCaptureExtension.class)
     @DirtiesContext
     @Test
-    void authChallengeTimeout(CapturedOutput output) {
+    void partialNonceReject(CapturedOutput output) {
         testLogOutput(
                 List.of(
                         "n.e.f.s.SSLCertificateCreateRenewService : Starting certificate create/renew",
                         "n.e.f.service.AcmeAccountService         : Account does not exist. Creating account.",
                         "n.e.f.s.SSLCertificateCreateRenewService : Certificate account login accessed",
                         "n.e.f.s.SSLCertificateCreateRenewService : Beginning certificate order. Renewal: false",
-                        "n.e.f.l.ChallengeTokenRequestedListener  : Timeout while checking for challenge status",
-                        "n.e.f.s.SSLCertificateCreateRenewService : Exception while ordering certificate, retry in 1 hours",
-                        "net.eightlives.friendlyssl.exception.SSLCertificateException: Exception while handling SSL certificate management",
-                        "Caused by: java.util.concurrent.TimeoutException: null"
+                        "n.e.f.service.UpdateCheckerService       : Resource is valid",
+                        "n.e.f.s.SSLCertificateCreateRenewService : Certificate renewal successful. New certificate expiration time is"
                 ),
                 output
         );
+
+        assertTrue(Files.exists(Path.of(config.getKeystoreFile())));
     }
 }
-
