+++ src/test/java/net/eightlives/friendlyssl/integration/tests/SlowServerTest.java
@@ -1,5 +1,7 @@
-package net.eightlives.friendlyssl.integration;
+package net.eightlives.friendlyssl.integration.tests;
 
+import net.eightlives.friendlyssl.integration.IntegrationTest;
+import net.eightlives.friendlyssl.integration.TestApp;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
@@ -10,22 +12,25 @@ import org.springframework.boot.test.system.OutputCaptureExtension;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
+import org.testcontainers.Testcontainers;
 import org.testcontainers.containers.BindMode;
 import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.junit.jupiter.Container;
 
 import java.util.List;
 
-@SpringBootTest
-@ContextConfiguration(initializers = TosNotAcceptedTest.class, classes = TestApp.class)
-@ActiveProfiles({"test-base", "test-tos-unaccepted"})
-public class TosNotAcceptedTest implements IntegrationTest {
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@ContextConfiguration(initializers = SlowServerTest.class, classes = TestApp.class)
+@ActiveProfiles({"test-base", "test-tos-accepted"})
+class SlowServerTest implements IntegrationTest {
 
-    @Container
+    static {
+        Testcontainers.exposeHostPorts(5002);
+    }
+
     static GenericContainer pebbleContainer = new GenericContainer("letsencrypt/pebble")
             .withCommand("pebble -config /test/my-pebble-config.json")
             .withExposedPorts(14000, 15000)
-            .withEnv("PEBBLE_VA_NOSLEEP", "1")
+            .withEnv("PEBBLE_VA_NOSLEEP", "0")
             .withEnv("PEBBLE_WFE_NONCEREJECT", "0")
             .withClasspathResourceMapping(
                     "pebble-config.json",
@@ -38,21 +43,22 @@ public class TosNotAcceptedTest implements IntegrationTest {
         return pebbleContainer;
     }
 
-    @DisplayName("Start server with no accepted terms of service")
+    @DisplayName("Start server and ACME server responses have variable latency")
     @Timeout(20)
     @ExtendWith(OutputCaptureExtension.class)
     @DirtiesContext
     @Test
-    void tosNotAccepted(CapturedOutput output) {
+    void slowServer(CapturedOutput output) {
         testLogOutput(
                 List.of(
                         "n.e.f.s.SSLCertificateCreateRenewService : Starting certificate create/renew",
-                        "n.e.f.service.AcmeAccountService         : Account does not exist. Terms of service must be accepted in file src/test/resources/integration/tos_unaccepted before account can be created",
-                        "n.e.f.s.SSLCertificateCreateRenewService : Exception while ordering certificate, retry in 1 hours",
-                        "net.eightlives.friendlyssl.exception.SSLCertificateException: Exception while handling SSL certificate management"
+                        "n.e.f.service.AcmeAccountService         : Account does not exist. Creating account.",
+                        "n.e.f.s.SSLCertificateCreateRenewService : Certificate account login accessed",
+                        "n.e.f.s.SSLCertificateCreateRenewService : Beginning certificate order. Renewal: false",
+                        "n.e.f.service.UpdateCheckerService       : Resource is valid",
+                        "n.e.f.s.SSLCertificateCreateRenewService : Certificate renewal successful. New certificate expiration time is"
                 ),
                 output
         );
     }
 }
-
